+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/TestOpMode.java
@@ -8,7 +8,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.wagner.GlobalStorage;
 
-@Autonomous(name = "Test OpMode")
+//@Autonomous(name = "Test OpMode")
 public class TestOpMode extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {